+++ vite.config.ts
@@ -11,6 +11,7 @@ export default defineConfig({
     dts({
       insertTypesEntry: true,
       include: ['src'],
+      exclude: ['src/**/*.test.*', 'src/**/*.spec.*'],
     }),
   ],
   build: {